Update project structure to be usable via HACS
