--- const.py
@@ -1,6 +1,0 @@
-DOMAIN = "olarm_sensors"
-CONF_UPDATE_INTERVAL = 5
-DEFAULT_UPDATE_INTERVAL = 5
-MIN_UPDATE_INTERVAL = 5
-AuthenticationError = "Invalid Credentials"
-ZONE = 0
